refactor: moved svg icons to shared folder
