+++ src/app/(store)/product/[productId]/page.tsx
@@ -6,11 +6,11 @@ import { useParams, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 
 import { getOneProduct } from "@/actions/product/product";
-import { LikeIcon, MinusIcon } from "@/components/icons/svgIcons";
 import ProductCard from "@/components/store/common/productCard";
 import Gallery from "@/components/store/productPage/gallery";
 import ProductBoard from "@/components/store/productPage/productBoard";
 import { TopProducts } from "@/domains/product/constants";
+import { LikeIcon, MinusIcon } from "@/shared/components/icons/svgIcons";
 import { SK_Box } from "@/shared/components/UI/skeleton";
 import { TProductPageInfo } from "@/types/product";
