+++ src/components/store/navbar/navShopping/index.tsx
@@ -3,12 +3,10 @@
 import { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { ShoppingIconOutline } from "@/components/icons/svgIcons";
+import { ShoppingIconOutline } from "@/shared/components/icons/svgIcons";
 import { cn } from "@/shared/utils/styling";
 import { TCartState, RootState } from "@/store/shoppingCart";
 import { toggleCart } from "@/store/shoppingCart";
-
-
 
 import ShoppingCart from "../../common/shoppingCart";
