+++ src/shared/components/icons/svgIcons.tsx
